feat(api-helper): refactor and move transport-related functions to transportHtml.ts
